+++ src/meltano/__init__.py
@@ -3,4 +3,4 @@
 from __future__ import annotations
 
 # Managed by commitizen
-__version__ = "3.5.2"
+__version__ = "3.5.4"