+++ tests/meltano/core/state_store/test_filesystem.py
@@ -11,11 +11,14 @@ import platform
 import shutil
 import string
 import typing as t
+import uuid
 from base64 import b64encode
 from contextlib import contextmanager
 from pathlib import Path
 from unittest.mock import MagicMock, PropertyMock, patch
 
+import botocore.exceptions
+import moto
 import pytest
 import time_machine
 from azure.core.exceptions import ResourceNotFoundError
@@ -439,6 +442,60 @@ class TestS3StateStoreManager:
 
     def test_state_path(self, subject: S3StateStoreManager) -> None:
         assert subject.state_dir == "state"
+
+    def test_set_first_time(self, monkeypatch: pytest.MonkeyPatch) -> None:
+        monkeypatch.delenv("AWS_DEFAULT_REGION", raising=False)
+        monkeypatch.delenv("AWS_PROFILE", raising=False)
+        state_id = uuid.uuid4().hex
+        with moto.mock_aws():
+            store_manager = S3StateStoreManager(
+                uri="s3://test_access_key_id:test_secret_access_key@meltano/state",
+                lock_timeout_seconds=10,
+            )
+            store_manager.client.create_bucket(Bucket=store_manager.bucket)
+            store_manager.set(JobState(state_id=state_id, completed_state={}))
+
+    def test_set_fail_object_in_glacier(
+        self,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        monkeypatch.delenv("AWS_DEFAULT_REGION", raising=False)
+        monkeypatch.delenv("AWS_PROFILE", raising=False)
+        state_id = uuid.uuid4().hex
+        with moto.mock_aws():
+            store_manager = S3StateStoreManager(
+                uri="s3://test_access_key_id:test_secret_access_key@meltano/state",
+                lock_timeout_seconds=10,
+            )
+            store_manager.client.create_bucket(Bucket=store_manager.bucket)
+            store_manager.client.put_object(
+                Bucket=store_manager.bucket,
+                Key=f"state/{state_id}/state.json",
+                Body=json.dumps({}).encode(),
+                StorageClass="GLACIER",
+            )
+            with pytest.raises(OSError, match="unable to access") as exc_info:
+                store_manager.set(JobState(state_id=state_id, completed_state={}))
+
+            exc = exc_info.value
+            assert isinstance(exc.__cause__, botocore.exceptions.ClientError)
+            assert exc.__cause__.response["Error"]["Code"] == "InvalidObjectState"
+
+    def test_set_fail_bucket_does_not_exist(
+        self,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        monkeypatch.delenv("AWS_DEFAULT_REGION", raising=False)
+        monkeypatch.delenv("AWS_PROFILE", raising=False)
+        with moto.mock_aws():
+            store_manager = S3StateStoreManager(
+                uri="s3://test_access_key_id:test_secret_access_key@meltano/state",
+                lock_timeout_seconds=10,
+            )
+            with pytest.raises(botocore.exceptions.ClientError) as exc_info:
+                store_manager.set(JobState(state_id="state-id", completed_state={}))
+
+            assert exc_info.value.response["Error"]["Code"] == "NoSuchBucket"
 
     def test_delete(self, subject: S3StateStoreManager) -> None:
         response = {